love-release is now modular
I have removed the platform-specific scripts form the main script an I
intend to store them in a specific location where they can be sourced by
the main script.
The mechanism to add a subscript is fairly simple. It is all explained
with comments inside the script.

I also moved out form the main script the getoptex function for
clarification into an include directory that will be use to store an
eventual INI parser later.